Fix TQ dequant MSE paged custom op build

+++ csrc/tq_dequant_mse_paged/op_host/tq_dequant_mse_paged_def.cpp
@@ -15,7 +15,7 @@ public:
     {
         this->Input("packedIdx")
             .ParamType(REQUIRED)
-            .DataType({ge::DT_FLOAT})
+            .DataType({ge::DT_UINT8})
             .Format({ge::FORMAT_ND})
             .AutoContiguous();
         this->Input("norm")
